feature/data-rest

+++ data-rest/liquibase/migrations/1.0.8/rollback_tadmingroup-default-records.sql
@@ -1,4 +1,1 @@
-delete from
-    tadmingroup
-where
-    group_name = "Assign permissions";
+delete from tadmingroup where group_name = "Privileged users";
